requirements.txt
asgiref==3.8.1
distlib==0.3.8
Django==5.0.7
django-cors-headers==4.4.0
djangorestframework==3.15.2
filelock==3.15.4
platformdirs==4.2.2
sqlparse==0.5.0
virtualenv==20.26.3
